Logic to control frontend display being battle-aware
When people collide, they start a battle.
When a battle starts, a player making a request
automatically gets shown the battle info.
When a player gets a request they get informed
if the output is of a zone or a battle and
displays appropriately.
Changed gls (game logic server) <-> web server api
to be a csv instead of line break delimited.